requirements.txt
openai
requests
bs4
telethon
pillow
aiohttp
ddgs
streamlit
python-dotenv
streamlit_authenticator